+++ packages/shared/friends/sagas.ts
@@ -138,8 +138,7 @@ import {
   isNewFriendRequestEnabled,
   decodeFriendRequestId,
   validateFriendRequestId,
-  reachedMaxNumberOfRequests,
-  isRemainingCooldown
+  DEFAULT_MAX_NUMBER_OF_REQUESTS
 } from './utils'
 import { AuthChain } from '@dcl/kernel-interface/dist/dcl-crypto'
 import { mutePlayers, unmutePlayers } from 'shared/social/actions'
@@ -2530,4 +2529,43 @@ function buildFriendRequestErrorResponse(error: FriendshipErrorCode) {
  */
 function buildFriendRequestReply<T>(reply: NonNullable<T>) {
   return { reply, error: undefined }
+}
+
+/**
+ * Check whether the user has reached the max number of sent requests to a given user
+ * @param userId
+ * @returns true if the user has reached the max number of sent requests, false otherwise
+ */
+export function reachedMaxNumberOfRequests(userId: string) {
+  // Get number friend requests sent in a session to the given user
+  const sentRequests = getNumberOfFriendRequests(store.getState())
+  const number = sentRequests.get(userId) ?? 0
+
+  // TODO Juli: Get max number allowed from config or ff
+  const maxNumber = DEFAULT_MAX_NUMBER_OF_REQUESTS
+
+  // Check current number vs max number allowed
+  if (number < maxNumber) {
+    return false
+  }
+  return true
+}
+
+/**
+ * Check whether there is a remaining cooldown time to send a friend request to a given user.
+ * @param userId
+ * @returns true if there is a remaining cooldown time and it hasn't expired yet, false otherwise
+ */
+export function isRemainingCooldown(userId: string) {
+  const currentTime = Date.now()
+
+  // Get the remaining cooldown time for the given user
+  const coolDownTimer = getCoolDownOfFriendRequests(store.getState())
+  const remainingCooldownTime = coolDownTimer.get(userId)
+
+  // If there is a remaining cooldown time and it hasn't expired yet, return false
+  if (remainingCooldownTime && currentTime < remainingCooldownTime) {
+    return true
+  }
+  return false
 }
